Adds deserialization method to Helper class
