+++ src/main/java/com/solstice/stocks/StockSummaryController.java
@@ -1,5 +1,6 @@
 package com.solstice.stocks;
 
+import com.solstice.stocks.data.StockQuoteRepository;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
