+++ src/main/java/com/solstice/stocks/data/StockQuoteRepository.java
@@ -1,5 +1,7 @@
-package com.solstice.stocks;
+package com.solstice.stocks.data;
 
+import com.solstice.stocks.model.StockQuote;
+import com.solstice.stocks.model.StockSummary;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
@@ -8,6 +10,7 @@ import java.util.List;
 
 @Repository
 public interface StockQuoteRepository extends CrudRepository<StockQuote, Long> {
+
 
     List<StockQuote> findAllBySymbol(String symbol);
 
@@ -18,4 +21,5 @@ public interface StockQuoteRepository extends CrudRepository<StockQuote, Long> {
     @Query(nativeQuery = true)
     StockSummary monthlySummaryQuery(@Param("symbol_in") String symbol_in,
                                     @Param("year_in") int year_in, @Param("month_in") int month_in);
+
 }
