+++ src/main/java/com/solstice/stocks/model/StockSummary.java
@@ -1,4 +1,4 @@
-package com.solstice.stocks;
+package com.solstice.stocks.model;
 
 import java.math.BigDecimal;
